use features directory
