+++ src/features/layout/Layout.tsx
@@ -2,9 +2,9 @@ import { GithubFilled, MenuOutlined } from "@ant-design/icons"
 import { Button, ConfigProvider, Drawer } from "antd"
 import { Link } from "gatsby"
 import * as React from "react"
-import '../styles/global.css'
-import Navigation from "./Navigation"
+import '../../styles/global.css'
 import * as styles from "./layout.module.css"
+import Navigation from "./Navigation"
 
 interface LayoutProps {
   children: React.ReactNode;
